controller/requirements.txt
textual==6.4.0
textual-dev==1.8.0
websocket-client==1.9.0